Phase 2 (D19): de-alias 125 CrystalDust trainers; restore rematch table
Phase 1 unified 125 CrystalDust trainers onto same-named Emerald trainers,
and left CrystalDust's 24 phone-rematch entries in a header nothing included.
Both are restored: new _GSC trainer IDs with 165 references retargeted, and
the rematch entries merged into expansion's REMATCH_* enum and gRematchTable.

Also fixes CrystalDust's bare-name MAP_NUM/MAP_GROUP call sites.

Errors 355 -> 261.

+++ include/phone_contact.h
@@ -20,8 +20,8 @@ struct PhoneContact
     u8 mapNum;
     u8 mapGroup;
     u16 registeredFlag;
-    u8 rematchTrainerId;
-    u8 trainerId;
+    u16 rematchTrainerId;
+    u16 trainerId;
     u8 availability;
     u8 isPermanent;
 };
